config: Default to using ~/.wgu/wgu.conf.
The existing logic requires the user to specify the configuration
file path as the first argument to "up". In the interest of making
wgu easier to use, this commit allows users to not specify a config
file. In such a case, wgu defaults to using the configuration file
at "~/.wgu/wgu.conf".