config: Created app settings section, replace "-A".
This commit introduces a new application / general settings
section to the configuration file. I decided to name it "wgu"
to avoid possible conflicts with WireGuard (i.e., "App" or
"General" would of been preferred... but I am afraid WireGuard
may use those some day). Maybe I will change the name later.

With this commit, the "-A" argument is replaced with the
"AutomaticAddressPlanningMode" parameter which is now
specified in the new configuration section.